@@ -124,7 +124,7 @@ struct KdTree
 
 		return ids;
 }
-	
+
 
 };
 
